# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend da POC do centralizador de saldo. Toda regra temporal usa POC_REFERENCE_DATE.",
        "title": "Centralizador de Saldo (POC) — API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/users/{userID}/accounts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Lista contas de um cliente com saldo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_httpapi.AccountDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}/assistant/plan": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assistant"
                ],
                "summary": "Turno do assistente de planejamento (chat)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Histórico da conversa",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.AssistantPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.AssistantPlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_httpapi.AccountDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "balance": {
                    "type": "string",
                    "example": "1234.56"
                },
                "balance_reference_date": {
                    "type": "string",
                    "example": "2025-01-01"
                },
                "branch_code": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                },
                "check_digit": {
                    "type": "string"
                },
                "compe_code": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_httpapi.AssistantPlanRequest": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.ChatMessageDTO"
                    }
                }
            }
        },
        "internal_httpapi.AssistantPlanResponse": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean"
                },
                "proposal": {
                    "$ref": "#/definitions/internal_httpapi.PlanProposalDTO"
                },
                "reply": {
                    "type": "string"
                }
            }
        },
...
        "internal_httpapi.ChatMessageDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Quero juntar 10 mil para uma viagem"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
        "internal_httpapi.PlanOptionDTO": {
            "type": "object",
            "properties": {
                "duration_months": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.PlanProposalDTO": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.ProposedAllocationDTO"
                    }
                },
                "name": {
                    "type": "string"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.PlanOptionDTO"
                    }
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.ProposedAllocationDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "percentage": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
    }
}